+++ src/features/albums/index.ts
@@ -1,1 +1,3 @@
 export { Albums } from './components/Albums';
+export { useFetchAlbums } from './hooks/useFetchAlbums';
+export type { Albums as AlbumType } from './types/albums-type';
